move funtions to util.py
