+++ bin/j1939_request_pgn.py
@@ -11,47 +11,6 @@ title = "%s Version: %s %s %s" % (_name, __version__, __date__, __exp__)
 import j1939
 
 
-def request_pgn_single(requested_pgn, channel='can0', bustype='socketcan', length=4, src=0, dest=0x17):
-
-    countdown = 10
-    result = None
-
-    if not isinstance(requested_pgn, int):
-        raise ValueError("pgn must be an integer.")
-
-    bus = j1939.Bus(channel=channel, bustype=bustype, timeout=0.01)
-    pgn = j1939.PGN()
-    pgn.value = 0xea00 + dest # request_pgn mem-object
-    aid = j1939.ArbitrationID(pgn=pgn, source_address=src, destination_address=dest)
-
-    pgn0 = requested_pgn & 0xff
-    pgn1 = (requested_pgn >> 8) & 0xff
-    pgn2 = (requested_pgn >> 16) & 0xff
-
-    data = [pgn0, pgn1, pgn2]
-    pdu = j1939.PDU(timestamp=0.0, arbitration_id=aid, data=data, info_strings=None)
-
-    pdu.display_radix='hex'
-
-    bus.send(pdu)
-
-    while countdown:
-        pdu = bus.recv(timeout=1)
-        if pdu and (pdu.pgn == 0xe800 or pdu.pgn == requested_pgn):
-            result = list(pdu.data) 
-            break # got what I was waiting for
-
-        if pdu: 
-            countdown -= 1
-
-    bus.shutdown()
-
-    if not result:
-        raise IOError(" no CAN response")
-
-
-    return result
-
 
 if __name__ == "__main__":
 
@@ -101,7 +60,7 @@ if __name__ == "__main__":
 
     print ("request_pgn_single(pgn=0x%04x (%d), src=0x%02x, dest=0x%02x)" % (pgn, pgn, source, dest))
 
-    val = request_pgn_single(pgn, length=4, src=source, dest=dest)
+    val = j1939.utils.request_pgn_single(pgn, length=4, src=source, dest=dest)
 
     print("returned PGN = %s" % val)
